Moved messages to own helper
